Move commont test functionality from RelationalExpressionTest.cpp to SparqlTestHelpers.cpp
